# BACKEND/src/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger AWS - Documentation",
    "description": "Esta es la documentación de la API de Swagger AWS donde se puede ver la información de los endpoints y los modelos de datos que se utilizan en la API asi como tambien conocer la estructura de los endpoints y los parametros que se utilizan en cada uno de ellos.", 
    "version": "1.0.14"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local server"
    },
    {
      "url": "https://apiaws.gobdigital.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/tenants/{tenantUUID}": {
      "get": {
        "tags": [
          "Tenant"
        ],
        "summary": "Obtener un tenant por su UUID",
        "description": "Obtener un tenant por su UUID guardado en la base de datos",
        "operationId": "getTenantByUUID",
        "parameters": [
          {
            "name": "tenantUUID",
            "in": "path",
            "description": "UUID del tenant que se quiere obtener",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "uuid": "oo58dJPPYbhRAEj7mxtXCo",
                  "nombre": "Nomifacil_014",
                  "presignedURLTime": 3000
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "uuid": {
                      "type": "string"
                    },
                    "nombre": {
                      "type": "string"
                    },
                    "presignedURLTime": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tenant no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 404,
                  "message": "Tenant no encontrado (AWS)"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Tenant"],
        "summary": "Actualizar un tenant por su UUID",
        "description": "Actualizar un tenant con su UUID",
        "operationId": "updateTenantByUUID",
        "parameters": [
          {
            "name": "nombre",
            "in": "body",
            "description": "Nombre que se le va a asignar al Tenant",
            "required": false,
            "schema": {
              "type": "string",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "NomiFacil_Euro_Updated"
                }
              }
            }
          },
          {
            "name": "presignedURLTime",
            "in": "body",
            "description": "Tiempo de expiración de la URL",
            "required": false,
            "schema": {
              "type": "number",
              "properties": {
                "presignedURLTime": {
                  "type": "integer",
                  "example": 5000
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consulta exitosa",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "uuid": "oo58dJPPYbhRAEj7mxtXCo",
                  "nombre": "NomiFacil_Euro_Updated",
                  "presignedURLTime": 5000
                }
              }
            }
          },
          "404": {
            "description": "Tenant no encontrado",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 404,
                  "message": "Tenant no encontrado (AWS)"
                }
              }
            }
          }
        }
      }
    },
...
    "/aws/{folder}/{fileName}": {
      "get": {
        "tags": ["AWS"],
        "summary": "Obtener un tenant por su UUID",
        "description": "Obtener un tenant por su UUID guardado en la base de datos",
        "operationId": "getFileByUUID",
        "parameters": [
          {
            "name": "folder",
            "in": "path",
            "description": "Nombre del folder donde se encuentra el archivo",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileName",
            "in": "path",
            "description": "Nombre del archivo",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpStatusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "requestId": {
                      "type": "string",
                      "example": "2KM4HGJV41W2RFJ0"
                    },
                    "extendedRequestId": {
                      "type": "string",
                      "example": "TaUWvgdtAsVAmAAKMOQ9eULdOggM0GDdnMpi51QuVR+yjzVM1e68drIpQy7CcuwSBu1nMK242b0="
                    },
                    "attempts": {
                      "type": "integer",
                      "example": 1
                    },
                    "totalRetryDelay": {
                      "type": "integer",
                      "example": 0
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Error al obtener un archivo cuya KEY no existe",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 404,
                  "message": "No se encontró el archivo"
                }
              }
            }
          }
        }        
      },
      "delete": {
        "tags": ["AWS"],
        "summary": "Eliminar un archivo",
        "description": "Eliminar un archivo del bucket de AWS S3 por su nombre y carpeta",
        "operationId": "deleteAWSFile",
        "parameters": [
          {
            "name": "folder",
            "in": "path",
            "description": "Nombre del folder donde se encuentra el archivo",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileName",
            "in": "path",
            "description": "Nombre del archivo",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archivo eliminado con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "$metadata": {
                      "type": "object",
                      "properties": {
                        "httpStatusCode": {
                          "type": "integer",
                          "example": 204
                        },
                        "requestId": {
                          "type": "string",
                          "example": "VD0WWTHS724TASXR"
                        },
                        "extendedRequestId": {
                          "type": "string",
                          "example": "ny7k8rmRjH6FBYxHVPuxIJrfBX/u0bRvzQhfYqOWnPD/1AZvp+PLCx3JEX+gnzzmfQfI0T8njAg="
                        },
                        "attempts": {
                          "type": "integer",
                          "example": 1
                        },
                        "totalRetryDelay": {
                          "type": "integer",
                          "example": 0
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
        
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}